Web Scraping Implemented
When the link the user searches doesn't exist in the database, app now scrapes web to get the objective fanfic data for that link and puts it in the database. User can immediately add review for it thereafter.

Still need to account for fanfic tags and possibly displaying multiple fandoms.