Tags Now Fetched With Fanfic
Updated the ao3 scraper to account for announcements.

The ao3 scraper now gets the tags associated with the work. The user can select which of those tags they liked. All of this is stored in the database. (Still need to update user list view to include tags.)

Added a class to create multi-select rectangles to display tags for the user to pick.

Removed dummy data inserts from database file.

Added a base url check before proceeding with scraping.